uploading: extract some nested classes from B2LargeFileUploader
...so they'll be more testable & available for small file uploading.

tested with unit tests ("./gradlew build javadoc")
exercised with B2Sample and watched output.
